backend/requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
djangorestframework-simplejwt==5.3.0

# Database - Using psycopg2 instead of psycopg2-binary for better compatibility
psycopg2==2.9.9
# prisma==0.12.0  # Commented out - compatibility issues with newer Python versions

# Authentication and Security
cryptography==41.0.7
PyJWT==2.8.0

# Utilities and Image Processing
python-decouple==3.8
Pillow==10.0.1
python-dotenv==1.0.0

# Development and Debugging
django-debug-toolbar==4.2.0
drf-spectacular==0.26.5

# HTTP and API
requests==2.31.0
urllib3==2.0.7

# Date and Time
python-dateutil==2.8.2

# Environment and Configuration
gunicorn==23.0.0

# Static Files (Production)
whitenoise==6.6.0

# Additional Dependencies
asgiref==3.7.2
sqlparse==0.4.4

# WebSocket Support
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1

# Excel Support
openpyxl==3.1.2

# Security and Dependency Scanning - Updated for better compatibility
safety==3.0.1
# pip-audit==2.6.1  # Commented out - may cause build issues
# bandit==1.7.5     # Commented out - may cause build issues

# Development Tools - Updated versions
black==24.1.1
flake8==7.0.0
isort==5.13.2
 
 